Change voice token caching mechanism to avoid 'Cannot set voice' error when selecting voice in C#. (Found in #16)
Previously, we cache the whole CVoiceToken and its CVoiceKey COM objects. Now we cache only the data strings necessary for constructing the tokens and keys (in a DataKeyData tree struct), and construct the corresponding COM objects only when necessary.

We use one shared_ptr for each CVoiceToken. All the token's CVoiceKey's share the data block with their root CVoiceToken. The shared data block will only be released when neither CVoiceToken nor its subkeys exist.

When the cache is still valid, we create brand new COM CVoiceToken's for the client, but they will share the same data blocks already in the cache.