Fix words gluing together when using JAWS (mentioned in #8), by adding spaces between text fragments when needed
SAPI separates text into fragments, and the text in different fragments must be different words.
Because of the XML parsing process, some spaces between different fragments may be removed. Unfortunately we have no idea what the original XML is, only the parsed fragments, so we have to take care of not letting adjacent fragments gluing into one word.

XML tags can be used to separate words, even when no spaces are around the tags. However, Edge voices have little support for XML tags, and removing the XML tags can cause words to be glued together. Even the supported tags such as <prosody> cannot be used more than a few times. So we have no choice but to insert spaces ourselves to separate the words.

Another caveat when using Chinese voices is that spaces add reading pauses between the words. So we should not add spaces in Chinese text. But there can be English words inside Chinese text, and we need spaces between English words. So we have to detect the language of the text and decide whether to insert a space or not accordingly.