Fix the occasional crash caused by writing audio outside Speak.
Most of the time, Speak will wait for the speaking to actually complete. But sometimes there can still be audio writing after Speak has returned, crashing the client application.

We fix this by resetting m_pOutputSite to null after Speak returns, and check the m_pOutputSite before using it. As there can be multiple threads, we use a mutex to protect it.