Add support for voice age properties
This is said to be "required" by SAPI's spec, although most of the time it is optional.

The problem is that "Age" is not in the voice list JSON. We have to get it from somewhere else, for example, from the Speech Studio webpage. To make it simple, we put a JSON file with some extra voice attributes in the resource data.

SAPI requires Age to be one of: Child, Teen, Adult, Senior. But the Azure voices uses: Child, YoungAdult, Adult, Senior (and unspecified). For now, the Azure age setting is used as-is.