Add connection pool to reduce the delay for online voices.
The connection pool keeps a few connections to the same host, and uses the connections in a round-robin manner. This way we don't have to close and wait for the connection to re-open if the speaking is stopped mid-way. Instead, we can just pick another connection in the pool.

The connection pool is process-wide, meaning that multiple instances in the same process can share the connection pool. For example, when using XML tags to switch between voices, SAPI will create multiple temporary instances of voice tokens, and now they can share the same connections.

The connection pool is divided into "sub-pools" by different hosts, or an endpoint URL plus a key. Connections in the same sub-pool uses the same URL and key.

Edge server actively closes connections inactive for 30 seconds. We can send something, for example, empty speech configs, to keep the connection alive longer. However, such a connection still cannot live longer than about 4 minutes. So we check the connections we are keeping alive, and if a connection lives long enough, we create a new connection to replace it.