Update SpeechRestAPI to use connection pool
Also refactor the code so that MP3 data processing is no longer in a dedicate backgournd thread, but is started by std::async every time it is needed, and ends when its work is done.

Connection creation/failure is now handled in the connection pool. MP3 data queue becomes a dedicated class.