Change module path to the forwarder's path when installing on ARM64 systems
According to [this][1] and [this][2], if we want to change how `%MODULE%` in the rgs scripts get replaced, we can use `ATL::CRegObject::AddReplacement`, but that's not easy because we aren't using `ATL::CRegObject` directly. Instead, we usually call
`CAtlComModule::DllRegisterServer` and let it do all the work.

`CAtlComModule::DllRegisterServer` will walk the object map, and call every object's `UpdateRegistry` method.

Every class that uses the `DECLARE_REGISTRY_RESOURCEID` macro to specify its rgs script will have a default `UpdateRegistry` definition, which usually calls `CAtlModule::UpdateRegistryFromResource` and pass it the resource ID. Turns out that `UpdateRegistryFromResource` accepts an argument, an array of `ATL::_ATL_REGMAP_ENTRY` that defines the replacement key-value pairs and ends with a pair of NULLs. But the `DECLARE_REGISTRY_RESOURCEID` macro doesn't pass any regmap nor allow us to specify one.

So usually we'll have to write the `UpdateRegistry` ourselves instead of using the `DECLARE_REGISTRY_RESOURCEID` macro. Or, we can redefine the macro to include a regmap, which is what I chose here, because we'll use a single global regmap.

Also, in each rgs script, the variable `%MODULE%` is changed to `%MODULEPATH%` to avoid conflicts with the predefined `%MODULE%`.

[1]: https://learn.microsoft.com/cpp/atl/using-replaceable-parameters-the-registrar-s-preprocessor
[2]: https://www.codeguru.com/soap/replaceable-parameters-for-atl-registrar/