Update connection pool
Use a wrapper for the connection returned by TakeConnection, so that the connection user can set its message and close handlers, without changing the real handler of the underlying connection. This way the connection pool won't lose control of its connections.

It also prevents data races when two threads change the same handler, or one changes the handler when the other is calling the handler.