Add support for Sec-MS-GEC
Microsoft Edge read aloud server now requires Sec-MS-GEC token to be provided in the request path, or the connection will be rejected, if the user's IP address is from mainland China. This adds the token generation algorithm into the program, so that Edge voices can still work with a Chinese IP address.

The algorithm requires the local time to be correct (synchronized with the server time). We calculate the time difference between the server time and the local time if a connection is rejected, so that we might get the local time "fixed" when attempting to make the next connection.

The algorithm: https://github.com/rany2/edge-tts/issues/290#issuecomment-2464956570