Use SAPI5's built-in implementation of SpObjectToken
At first, I thought that both ISpDataKey and ISpObjectToken need to be implemented. Also, if the automation interfaces ISpeechDataKey and ISpeechObjectToken aren't implemented, languages such as VB won't be able to use those objects.

But it turns out that SAPI5 has a built-in implementation of SpObjectToken that can be initialized with ISpObjectTokenInit::InitFromDataKey, which accepts an ISpDataKey and the token's ID to turn the data key into an object token.

The good news is that we can provide our own implementation of ISpDataKey to it, and it can still work! Not only that, the object token implementation also handles all automation interfaces, so we don't have to write our own implementation. The only thing we need to provide is an ISpDataKey implementation.

The bad news is that, all my previous implementations of ObjectTokens and automation interfaces become unnecessary...