Prevent loading the enumerator inside Windows 11 Narrator process and SystemSettings.exe to prevent Speech SDK version collision
Narrator supporting natural voices will load its own version of Speech SDK.

But if this enumerator is loaded first, a different version of the SDK will be loaded, and the Narrator's version of SDK will not be loaded due to the same DLL names.

To avoid collision, we disable the enumerator when running inside the Narrator process.