Do not wait for cancellation to complete on requested stop
Local natural voices need some time to complete cancellation. Even if cancellation is requested immediately after speaking, it will take about 100ms to complete. This will force the client requesting stop to wait as well.

Screen readers like NVDA may change the text to be spoken frequently. If there's a lot of stopping and speaking requests in a short period of time, the extra cancellation delay during each stops may stuck the app for an extended period of time.

To address this issue, the cancellation is put in the background, and Speak() will return immediately after that. This way the client will not wait on stopping the speech. Before processing the next speech, we'll check if there's a previous cancellation in progress, and wait for that first - waiting during speaking will not block the client, just delaying the speech. If the current speech is stopped before the previous cancellation completes, we can ignore the stopped speech, and it will not introduce extra cancellation delay.